warning removed

+++ src/Net.cpp
@@ -67,7 +67,7 @@ void Net::classify_all(const MatrixFloat& mIn, MatrixFloat& mClass) const
     for(int i=0;i<mIn.rows();i++)
     {
         forward(mIn.row(i),mOut);
-        mClass(i,0)= argmax(mOut);
+        mClass(i,0)= (float)argmax(mOut);
     }
 }
 /////////////////////////////////////////////////////////////////////////////////////////////////
